docs(Extender): rename file, added docs
